requirements.txt
anthropic[vertex]>=0.42.0
python-telegram-bot>=21.0
tavily-python>=0.5.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0
pytest-asyncio>=0.24
